requirements.txt
aiohttp==3.8.1
aioredis==2.0.1
confluent_kafka==1.8.2
python-dotenv==0.20.0
python_dateutil==2.8.2
websockets==10.1
